Rename: bot directory to agent
