+++ .gitignore
@@ -118,3 +118,5 @@ dmypy.json
 
 # Pyre type checker
 .pyre/
+
+.idea
